Create jaxb-impl-2.3.0.jar
Java 9, 10 or 11, then you will encounter an error when you run your Hibernate program.
This happens because of Java 9 and higher.

Java 9 and higher has removed java.xml.bind from its default classpath. That's why we get the class not found exception.  We have to explicitly add JAR files to the build path.